refactored main.java to AccountingLedgerApplication, added ConsoleUtils, cleaned Transaction
